Populate the Bug Report button with passed-in variables
It's critical that we pass-along any variables we've received in the
URL so that the BSA can make full use of them. The BSA is not
currently able to process them (due to some iframe shenanigans), but
we hope to be able to very soon, which will make it MUCH easier for us
to know EXACTLY what version of LibreOffice a user is running (along
with a handful of other useful pieces of data!)

Here's an example of the variables LibreOffice passes-along:

  ?LOversion=4.1.2.3&
   LOlocale=en-US&
   LOmodule=TextDocument

Neat!